mcp-server/requirements.txt
mcp==1.0.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4